Разбил на 3 модуля js
