+++ js/util.js
@@ -1,0 +1,7 @@
+const getRandomPositiveInteger = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min)) + min;
+};
+export {getRandomPositiveInteger};
+
